requirements.txt
selenium
webdriver-manager
selenium-stealth
python-dotenv
bs4
requests
